create separate module for preprocessing steps
