+++ src/vame/preprocessing/align_egocentrical.py
